Rename spatial to distance
